feat: add Show Sessions Picker action and refactor session actions
