+++ src/vs/sessions/contrib/sessions/browser/sessions.contribution.ts
@@ -16,6 +16,7 @@ import { SessionsManagementService, ISessionsManagementService } from './session
 import { SessionsTitleBarContribution } from './sessionsTitleBarWidget.js';
 import { SessionsView, SessionsViewId } from './views/sessionsView.js';
 import './views/sessionsViewActions.js';
+import './sessionsActions.js';
 
 const agentSessionsViewIcon = registerIcon('chat-sessions-icon', Codicon.commentDiscussionSparkle, localize('agentSessionsViewIcon', 'Icon for Agent Sessions View'));
 const AGENT_SESSIONS_VIEW_TITLE = localize2('agentSessions.view.label', "Sessions");
